requirements.txt
Django==2.1.5
django-ckeditor==5.6.1
django-cors-headers==2.2.0
django-js-asset==1.1.0
django-nested-admin==3.0.21
django-rest-framework-redocs==0.0.7
djangorestframework==3.8.2
Pillow==5.2.0
python-monkey-business==1.0.0
pytz==2018.5
six==1.11.0
Naked==0.1.31
oauth2client==4.1.2
pushjack==1.5.0
pyfcm==1.4.5
boto3
django-storages
psycopg2-binary
uWSGI==2.0.17
nexmo==2.3.0
django-bootstrap3==9.1.0
djangorestframework-jwt==1.11.0
